fix : remove all unser param/lib, remove token function to ultis
