+++ src/auth/auth.module.ts
@@ -5,11 +5,12 @@ import { AuthController } from './auth.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { UsersModule } from 'src/users/users.module';
 import { JWTStrategy } from './strategy';
+import { TokenService } from './ultis/token.service';
 
 @Module({
   imports: [JwtModule.register({}), forwardRef(() => UsersModule)],
   controllers: [AuthController],
-  providers: [AuthService, JWTStrategy],
+  providers: [AuthService, JWTStrategy, TokenService],
   exports: [AuthService],
 })
 export class AuthModule {}
